Improved COPR support in manage script #235
