+++ tools/update.py
@@ -19,48 +19,11 @@ import sys
 from l2tdevtools import download_helper
 from l2tdevtools import presets
 from l2tdevtools import projects
+from l2tdevtools import versions
 
 
 if platform.system() == 'Windows':
   import wmi  # pylint: disable=import-error
-
-
-def CompareVersions(first_version_list, second_version_list):
-  """Compares two lists containing version parts.
-
-  Note that the version parts can contain alpha numeric characters.
-
-  Args:
-    first_version_list (list[str]): first version parts.
-    second_version_list (list[str]): second version parts.
-
-  Returns:
-    int: 1 if the first is larger than the second, -1 if the first is smaller
-        than the second, or 0 if the first and second are equal.
-  """
-  first_version_list_length = len(first_version_list)
-  second_version_list_length = len(second_version_list)
-
-  for index in range(0, first_version_list_length):
-    if index >= second_version_list_length:
-      return 1
-
-    try:
-      first_version_part = int(first_version_list[index], 10)
-      second_version_part = int(second_version_list[index], 10)
-    except ValueError:
-      first_version_part = first_version_list[index]
-      second_version_part = second_version_list[index]
-
-    if first_version_part > second_version_part:
-      return 1
-    elif first_version_part < second_version_part:
-      return -1
-
-  if first_version_list_length < second_version_list_length:
-    return -1
-
-  return 0
 
 
 class GithubRepoDownloadHelper(download_helper.DownloadHelper):
@@ -408,7 +371,8 @@ class DependencyUpdater(object):
       if name not in package_versions:
         compare_result = 1
       else:
-        compare_result = CompareVersions(version, package_versions[name])
+        compare_result = versions.CompareVersions(
+            version, package_versions[name])
 
       if compare_result > 0:
         package_filenames[name] = package_filename
@@ -644,7 +608,7 @@ class DependencyUpdater(object):
             compare_result = -1
           else:
             version_tuple = version.split('.')
-            compare_result = CompareVersions(
+            compare_result = versions.CompareVersions(
                 version_tuple, package_versions[name])
             if compare_result >= 0:
               # The latest or newer version is already installed.
@@ -732,7 +696,7 @@ class DependencyUpdater(object):
         elif not found_package:
           compare_result = 1
         else:
-          compare_result = CompareVersions(
+          compare_result = versions.CompareVersions(
               version_tuple, package_versions[name])
           if compare_result >= 0:
             # The latest or newer version is already installed.
